client/pages/api/v1/transactions/user/[userId]/all.ts
import { NextApiRequest, NextApiResponse } from 'next';
import { getServerSession } from 'next-auth';

export default async function getAllTxnsByUserBetweenDatesHandler(
  req: NextApiRequest,
  res: NextApiResponse,
) {
  if (req.method !== 'GET') {
    return res.status(405).json({ error: 'invalid method' });
  }

  const session = await getServerSession();
  if (!session) {
    return res.status(401).json({ error: 'no valid session found' });
  }
  if (session.user?.email !== req.query.userId) {
    return res
      .status(403)
      .json({ error: "you cannot view another person's transactions" });
  }
}
